Fix fetch number (#16)
* Add `throwHttpErrors` to `fetchOrders`
When looking at `fetchOrders` we realised we needed to add in the `throwHttpErrors` option so we can handle non-2xx errors ourselves.

* Subtract 1 from fetch 'to' and 'end'values
We noticed that the end values of the "Fetching bundle" message were wrong, eg. if fetching 3 bundles it would display `(1-4/4)`. We subtract 1 from the 'to' value and the end value to fix this.